api defense

+++ app/api/admin/values/[id]/route.ts
@@ -1,7 +1,20 @@
+import { authConfig } from '@/configs/auth.config';
+import { userIsAdmin } from '@/services/orm/users.service';
 import { deleteValue, updateValue } from '@/services/orm/values.service';
+import { getServerSession } from 'next-auth';
 import { NextRequest, NextResponse } from 'next/server';
 
 export async function PUT (req: NextRequest, { params }: { params: Params }) {
+  const session = await getServerSession(authConfig);
+  
+  if (!session) {
+    return NextResponse.json({ code: 401, message: 'Unauthorized' });
+  }
+
+  if (!(await userIsAdmin({ where: { id: session.user.id } }))) {
+    return NextResponse.json({ code: 403, message: 'Forbidden' });
+  }
+
   const data = await req.json() as InputPut;
 
   const value = await updateValue(Number(params.id), data);
@@ -10,6 +23,16 @@ export async function PUT (req: NextRequest, { params }: { params: Params }) {
 }
 
 export async function DELETE (req: NextRequest, { params }: { params: Params }) {
+  const session = await getServerSession(authConfig);
+  
+  if (!session) {
+    return NextResponse.json({ code: 401, message: 'Unauthorized' });
+  }
+
+  if (!(await userIsAdmin({ where: { id: session.user.id } }))) {
+    return NextResponse.json({ code: 403, message: 'Forbidden' });
+  }
+
   await deleteValue(Number(params.id));
 
   return NextResponse.json({ code: 200 });
